Implement review list (#38)
* Update driver stats after a ride

* Implement endpoint for fetching driver reviews

* Implement driver review list

+++ spring-backend/src/main/java/com/example/springbackend/dto/display/DetailedRideHistoryPassengerDTO.java
@@ -7,6 +7,6 @@ import java.util.*;
 @Data
 public class DetailedRideHistoryPassengerDTO {
     private DriverDisplayDTO driver;
-    private Map<String, Double> driverRating = new HashMap<String, Double>() {};
-    private Map<String, Double> vehicleRating = new HashMap<String, Double>() {};
+    private Map<String, Integer> driverRating = new HashMap<String, Integer>() {};
+    private Map<String, Integer> vehicleRating = new HashMap<String, Integer>() {};
 }
